rework folder structure; move Chat.js to pages, create Home
