+++ src/pages/MapJ.js
@@ -1,0 +1,13 @@
+// Jade's map main page goes here
+
+import React from "react"
+
+class MapJ extends React.Component {
+    render() {
+        return (
+            <p>doot doot?</p>
+        )
+    }
+}
+
+export default MapJ;
