+++ src/pages/MapY.js
@@ -1,0 +1,1 @@
+// Yong's map main page goes here
